project to sub directories, GeoLocation to class
Re-organized the project to sub directories, changed GeoLocation from functions to a class
